Rename Main view
